requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cors==1.0.1
Django==5.1.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.16.1
future==1.0.0
gevent==24.10.3
greenlet==3.1.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
psycopg==3.2.3
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
rpds-py==0.20.0
setuptools==75.2.0
sqlparse==0.5.1
tldextract==5.1.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zope.event==5.0
zope.interface==7.1.0
